refactor: extract InvitationManager from models to services
Move InvitationManager + WorkshopInvitationManagerConcerns from
app/models/ to app/services/ as a single file, dissolving the
code-slicing concern.

This is a layer violation fix per Layered Design for Ruby on Rails
Applications (Dementyev, 2024):
- Domain layer (models) should not call mailers directly
- Application layer (services) is the correct home for orchestration
- Concerns should be behavioral, not group methods of one class

Specs moved from spec/models/ to spec/services/ accordingly.
