+++ spec/services/invitation_manager_spec.rb
